Fix gwtc (can only hash file plaintext up to 4 GiB)

+++ src/peergos/shared/crypto/hash/NativeScryptJS.java
@@ -18,5 +18,5 @@ public class NativeScryptJS {
 
     public native byte[] blake2b(byte[] input, int outputBytes);
 
-    public native CompletableFuture<byte[]> streamSha256(AsyncReader stream, long length);
+    public native CompletableFuture<byte[]> streamSha256(AsyncReader stream, int length);
 }
